+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/HandlerAdapter.java
@@ -1,8 +1,7 @@
-package webmvc.org.springframework.web.servlet.mvc;
+package webmvc.org.springframework.web.servlet;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.ModelAndView;
 
 public interface HandlerAdapter {
