+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/HandlerMapping.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc;
+package webmvc.org.springframework.web.servlet;
 
 import jakarta.servlet.http.HttpServletRequest;
