+++ mvc/src/test/java/webmvc/org/springframework/web/servlet/HandlerMappingRegistryTest.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc;
+package webmvc.org.springframework.web.servlet;
 
 import jakarta.servlet.http.HttpServletRequest;
 import org.junit.jupiter.api.DisplayName;
